+++ app/Commands/OneDrive/Download.php
@@ -6,6 +6,8 @@ use App\Helpers\OneDrive;
 use App\Helpers\Tool;
 use Illuminate\Console\Scheduling\Schedule;
 use LaravelZero\Framework\Commands\Command;
+use Symfony\Component\Process\Process;
+use Symfony\Component\Process\Exception\ProcessFailedException;
 
 class Download extends Command
 {
@@ -14,8 +16,10 @@ class Download extends Command
      *
      * @var string
      */
-    protected $signature = 'download {remote? : Remote Path}
-                                     {--id= : ID}';
+    protected $signature = 'download {local? : Download Local Path }
+                                     {remote? : Download Remote Path}
+                                     {--id= : Download Remote File ID}
+                                     {--hack : Download Via aria2c}';
 
     /**
      * The description of the command.
@@ -30,16 +34,33 @@ class Download extends Command
     public function handle()
     {
         $this->call('refresh:token');
-        if ($id = $this->option('id')) {
+        $remote = $this->argument('remote');
+        $local = $this->argument('local');
+        $id = $this->option('id');
+        $hack = $this->option('hack');
+        if (empty($local)) exit('Parameters Missing!');
+        if ($id) {
             $result = OneDrive::getItem($id);
         } else {
-            $remote = $this->argument('remote');
             if (empty($remote)) exit('Parameters Missing!');
             $graphPath = Tool::getRequestPath($remote);
             $result = OneDrive::getItemByPath($graphPath);
         }
         $response = Tool::handleResponse($result);
-        $response['code'] === 200 ? $this->info("Download Link:\n{$response['data']['@microsoft.graph.downloadUrl']}") : $this->warn("Failed!\n{$response['msg']} ");
+        if ($response['code'] === 200) {
+            $download = $response['data']['@microsoft.graph.downloadUrl'];
+            if (strtolower(PHP_OS) == "winnt") {
+                $this->warn("Download Not Support Windows");
+                $this->info("Download Link:\n{$download}");
+                exit;
+            }
+            $command = $hack ? "wget --no-check-certificate -c -O --tries=16 {$local} {$download}" : "aria2c -c -o {$local} -s16 -x16 -k1M {$download}";
+            $process = new Process($command);
+            $process->run();
+            if (!$process->isSuccessful()) {
+                throw new ProcessFailedException($process);
+            }
+        } else  $this->warn("Failed!\n{$response['msg']} ");
     }
 
     /**